java_code_inspector/requirements.txt
javalang>=0.13.0
pylint>=2.12.0
astunparse>=1.6.0
pandas>=1.3.0
jinja2>=3.0.0
xmltodict>=0.12.0